agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litestar[standard]",
    "langgraph",
    "langchain",
    "langchain-openai",
    "uvicorn",
    "python-multipart",
    "websockets",
    "pydantic",
    "python-dotenv",
]
